+++ core/src/main/java/org/multibit/hd/hardware/core/wallets/HardwareWallets.java
@@ -88,8 +88,8 @@ public class HardwareWallets {
    */
   public static <T extends HardwareWallet> T newUsbInstance(
     Class<T> hardwareWalletClass,
-    Optional<Integer> vendorId,
-    Optional<Integer> productId,
+    Optional<Short> vendorId,
+    Optional<Short> productId,
     Optional<String> serialNumber
   ) {
 
@@ -119,8 +119,8 @@ public class HardwareWallets {
    */
   public static HardwareWallet newUsbInstance(
     String className,
-    Optional<Integer> vendorId,
-    Optional<Integer> productId,
+    Optional<Short> vendorId,
+    Optional<Short> productId,
     Optional<String> serialNumber
   ) {
