+++ examples/src/main/java/org/multibit/hd/hardware/examples/trezor/usb/UsbSigningExample.java
@@ -64,8 +64,8 @@ public class UsbSigningExample {
 
     TrezorV1UsbHardwareWallet wallet = HardwareWallets.newUsbInstance(
       TrezorV1UsbHardwareWallet.class,
-      Optional.<Integer>absent(),
-      Optional.<Integer>absent(),
+      Optional.<Short>absent(),
+      Optional.<Short>absent(),
       Optional.<String>absent()
     );
